fix: display pretty evaluation errors for workflow evaluation unit tests.
